Refactor: refactor test
- move influx util to test/util/influx (ex: bucket recreation
- separate integration and end2end package
