+++ test/end2end/main_test.go
@@ -2,11 +2,11 @@ package integration
 
 import (
 	"context"
-	"fmt"
 	"os"
 	"testing"
 	"time"
 
+	influxutil "github.com/Goboolean/core-system.worker/test/util/influx"
 	influxdb2 "github.com/influxdata/influxdb-client-go/v2"
 	"github.com/influxdata/influxdb-client-go/v2/api/write"
 	"github.com/stretchr/testify/assert"
@@ -24,52 +24,6 @@ var (
 	annotationBucket = os.Getenv("INFLUXDB_ANNOTATION_BUCKET")
 )
 
-func RecreateBucket(client influxdb2.Client, orgName, bucketName string) error {
-	org, err := client.OrganizationsAPI().FindOrganizationByName(context.Background(), orgName)
-	if err != nil {
-		return err
-	}
-
-	bucket, err := client.BucketsAPI().FindBucketByName(context.Background(), bucketName)
-	if err != nil {
-		return nil
-	}
-
-	if err := client.BucketsAPI().DeleteBucket(context.Background(), bucket); err != nil {
-		return err
-	}
-	_, err = client.BucketsAPI().CreateBucketWithName(context.Background(), org, bucketName)
-
-	return err
-}
-
-func CountRecordsInMeasurement(client influxdb2.Client, orgName, bucketName, measurement string) (int, error) {
-
-	q, err := client.QueryAPI(orgName).
-		Query(context.Background(),
-			fmt.Sprintf(
-				`from(bucket: "%s")
-				|> range(start:0)
-				|> filter(fn: (r) => r["_measurement"] == "%s")
-				|> count()`, bucketName, measurement))
-	if err != nil {
-		return 0, err
-	}
-
-	num := int64(0)
-
-	// 각 record 별 count에서 최댓값을 찾는다.
-	for q.Next() {
-		fmt.Println("CountRecordsInMeasurement: records: ", q.Record().Values())
-		val := q.Record().ValueByKey("_value").(int64)
-		if val > num {
-			num = val
-		}
-	}
-
-	return int(num), nil
-}
-
 func TestPing(t *testing.T) {
 	ok, err := rawInfluxClient.Ping(context.Background())
 	assert.True(t, ok)
@@ -79,8 +33,7 @@ func TestPing(t *testing.T) {
 func TestCountRecordsInMeasurement(t *testing.T) {
 	bucket := annotationBucket
 	measurement := "testMeasurement"
-
-	if err := RecreateBucket(rawInfluxClient, influxDBOrg, bucket); err != nil {
+	if err := influxutil.RecreateBucket(rawInfluxClient, influxDBOrg, bucket); err != nil {
 		t.Error(err)
 		t.FailNow()
 	}
@@ -105,7 +58,7 @@ func TestCountRecordsInMeasurement(t *testing.T) {
 		}
 	}
 
-	count, err := CountRecordsInMeasurement(rawInfluxClient, influxDBOrg, bucket, measurement)
+	count, err := influxutil.CountRecordsInMeasurement(rawInfluxClient, influxDBOrg, bucket, measurement)
 	assert.Equal(t, num, count)
 	assert.NoError(t, err)
